+++ tests/testthat/test-derive_param_clinbenefit.R
@@ -1,10 +1,5 @@
-library(tibble)
-library(dplyr)
-library(lubridate)
-library(admiraldev)
-library(admiral)
 
-adsl <- tribble(
+adsl <- tibble::tribble(
   ~USUBJID, ~TRTSDT,      ~EOSDT,
   "01",     "2020-12-06", "2022-03-06",
   "02",     "2021-01-16", "2022-02-03",
@@ -20,7 +15,7 @@ adsl <- tribble(
     EOSDT = as_date(EOSDT),
   )
 
-adrs <- tribble(
+adrs <- tibble::tribble(
   ~USUBJID, ~PARAMCD, ~AVALC,          ~ADT,
   "01",     "RSP",    "Y",             "2021-04-08",
   "02",     "RSP",    "N",             "2021-05-07",
@@ -59,21 +54,21 @@ adrs <- tribble(
     new_vars = exprs(TRTSDT)
   )
 
-pd <- date_source(
+pd <- admiral::date_source(
   dataset_name = "adrs",
   date = ADT,
   filter = PARAMCD == "PD" & AVALC == "Y" & ANL01FL == "Y"
 )
 
-resp <- date_source(
+resp <- admiral::date_source(
   dataset_name = "adrs",
   date = ADT,
   filter = PARAMCD == "RSP" & AVALC == "Y" & ANL01FL == "Y"
 )
 
-# Clinical benefit rate parameter is derived correctly, Test 1 ----
-test_that("Clinical benefit rate parameter is derived correctly Test 1: ignore NON-CR/NON-PD", {
-  input_cbr <- tribble(
+## Test 1: ignore NON-CR/NON-PD ----
+test_that("derive_param_clinbenefit Test 1: ignore NON-CR/NON-PD", {
+  input_cbr <- tibble::tribble(
     ~USUBJID, ~PARAMCD, ~AVALC, ~AVAL, ~ADT,
     "01",     "CBR",    "Y",    1,     "2021-03-07",
     "02",     "CBR",    "Y",    1,     "2021-03-07",
@@ -104,6 +99,7 @@ test_that("Clinical benefit rate parameter is derived correctly Test 1: ignore N
     ref_start_window = 28,
     clinben_vals = c("CR", "PR", "SD"),
     set_values_to = exprs(
+      AVAL = yn_to_numeric(AVALC),
       PARAMCD = "CBR",
       ANL01FL = "Y"
     )
@@ -114,8 +110,8 @@ test_that("Clinical benefit rate parameter is derived correctly Test 1: ignore N
   )
 })
 
-# Clinical benefit rate parameter is derived correctly, Test 2 ----
-test_that("Clinical benefit rate parameter is derived correctly Test 2: No source_pd", {
+## Test 2: No source_pd ----
+test_that("derive_param_clinbenefit Test 2: No source_pd", {
   input_cbr <- tribble(
     ~USUBJID, ~PARAMCD, ~AVALC, ~AVAL, ~ADT,
     "01",     "CBR",    "Y",    1,     "2021-03-07",
@@ -146,6 +142,7 @@ test_that("Clinical benefit rate parameter is derived correctly Test 2: No sourc
     reference_date = TRTSDT,
     ref_start_window = 28,
     set_values_to = exprs(
+      AVAL = yn_to_numeric(AVALC),
       PARAMCD = "CBR",
       ANL01FL = "Y"
     )
@@ -154,5 +151,52 @@ test_that("Clinical benefit rate parameter is derived correctly Test 2: No sourc
   expect_dfs_equal(actual_output_no_source_pd,
     expected_output_no_source_pd,
     keys = c("USUBJID", "PARAMCD", "ADT")
+  )
+})
+
+## Test 3: Deprecation warning for aval_fun ----
+test_that("derive_param_clinbenefit Test 3: Deprecation warning for aval_fun", {
+  input_cbr <- tribble(
+    ~USUBJID, ~PARAMCD, ~AVALC, ~AVAL, ~ADT,
+    "01",     "CBR",    "Y",    1,     "2021-03-07",
+    "02",     "CBR",    "Y",    1,     "2021-03-07",
+    "03",     "CBR",    "N",    0,     NA,
+    "04",     "CBR",    "N",    0,     NA,
+    "05",     "CBR",    "N",    0,     NA,
+    "06",     "CBR",    "Y",    1,     "2021-09-22",
+    "07",     "CBR",    "Y",    1,     "2021-12-05"
+  ) %>%
+    mutate(
+      STUDYID = "AB42",
+      ADT = as_date(ADT),
+      ANL01FL = "Y"
+    ) %>%
+    left_join(adsl, by = c("STUDYID", "USUBJID")) %>%
+    select(-EOSDT)
+
+  expected_output_no_source_pd <- bind_rows(adrs, input_cbr)
+
+  expect_warning(
+    actual_output_no_source_pd <- derive_param_clinbenefit(
+    dataset = adrs,
+    dataset_adsl = adsl,
+    filter_source = PARAMCD == "OVR",
+    source_resp = resp,
+    source_pd = NULL,
+    source_datasets = list(adrs = adrs),
+    reference_date = TRTSDT,
+    ref_start_window = 28,
+    aval_fun = yn_to_numeric,
+    set_values_to = exprs(
+      PARAMCD = "CBR",
+      ANL01FL = "Y"
+    )
+  ),
+  class = "lifecycle_warning_deprecated"
+  )
+
+  expect_dfs_equal(actual_output_no_source_pd,
+                   expected_output_no_source_pd,
+                   keys = c("USUBJID", "PARAMCD", "ADT")
   )
 })