+++ tests/testthat/test-derive_param_confirmed_resp.R
@@ -1,10 +1,5 @@
-library(tibble)
-library(dplyr)
-library(lubridate)
-library(admiraldev)
-library(admiral)
 
-adsl <- tribble(
+adsl <- tibble::tribble(
   ~USUBJID, ~TRTSDTC,
   "1",      "2020-01-01",
   "2",      "2019-12-12",
@@ -20,7 +15,7 @@ adsl <- tribble(
     STUDYID = "XX1234"
   )
 
-adrs <- tribble(
+adrs <- tibble::tribble(
   ~USUBJID, ~ADTC,        ~AVALC,
   "1",      "2020-01-01", "PR",
   "1",      "2020-02-01", "CR",
@@ -53,17 +48,17 @@ adrs <- tribble(
   "9",      "2020-03-06", "SD"
 ) %>%
   mutate(PARAMCD = "OVR") %>%
-  bind_rows(tribble(
+  bind_rows(tibble::tribble(
     ~USUBJID, ~ADTC,        ~AVALC,
     "9",      "2020-02-16", "Y"
   ) %>%
     mutate(PARAMCD = "PD")) %>%
   mutate(
-    ADT = ymd(ADTC),
+    ADT = lubridate::ymd(ADTC),
     STUDYID = "XX1234"
   )
 
-pd_date <- date_source(
+pd_date <- admiral::date_source(
   dataset_name = "adrs",
   date = ADT,
   filter = PARAMCD == "PD"
@@ -82,6 +77,7 @@ test_that("derive_param_confirmed_resp Test 1: default confirmed response", {
         source_datasets = list(adrs = adrs),
         ref_confirm = 28,
         set_values_to = exprs(
+          AVAL = yn_to_numeric(AVALC),
           PARAMCD = "CRSP",
           PARAM = "Confirmed Response by Investigator"
         )
@@ -91,7 +87,7 @@ test_that("derive_param_confirmed_resp Test 1: default confirmed response", {
 
   expected <- bind_rows(
     adrs,
-    tribble(
+    tibble::tribble(
       ~USUBJID, ~ADTC,         ~AVALC, ~AVAL,
       "1",      "2020-01-01",  "Y",    1,
       "2",      NA_character_, "N",    0,
@@ -104,7 +100,7 @@ test_that("derive_param_confirmed_resp Test 1: default confirmed response", {
       "9",      NA_character_, "N",    0
     ) %>%
       mutate(
-        ADT = ymd(ADTC),
+        ADT = lubridate::ymd(ADTC),
         STUDYID = "XX1234",
         PARAMCD = "CRSP",
         PARAM = "Confirmed Response by Investigator"
@@ -122,13 +118,13 @@ test_that("derive_param_confirmed_resp Test 1: default confirmed response", {
 test_that("derive_param_confirmed_resp Test 2: accept SD", {
   adrs_ext <- bind_rows(
     filter(adrs, USUBJID != "7"),
-    tribble(
+    tibble::tribble(
       ~USUBJID, ~ADTC,        ~AVALC,
       "7",      "2020-04-02", "ND"
     ) %>%
       mutate(
         PARAMCD = "OVR",
-        ADT = ymd(ADTC),
+        ADT = lubridate::ymd(ADTC),
         STUDYID = "XX1234"
       )
   )
@@ -145,6 +141,7 @@ test_that("derive_param_confirmed_resp Test 2: accept SD", {
         max_nr_ne = 0,
         accept_sd = TRUE,
         set_values_to = exprs(
+          AVAL = yn_to_numeric(AVALC),
           PARAMCD = "CRSP",
           PARAM = "Confirmed Response by Investigator"
         )
@@ -154,7 +151,7 @@ test_that("derive_param_confirmed_resp Test 2: accept SD", {
 
   expected <- bind_rows(
     adrs_ext,
-    tribble(
+    tibble::tribble(
       ~USUBJID, ~ADTC,         ~AVALC, ~AVAL,
       "1",      "2020-01-01",  "Y",    1,
       "2",      "2020-02-01",  "Y",    1,
@@ -167,7 +164,7 @@ test_that("derive_param_confirmed_resp Test 2: accept SD", {
       "9",      NA_character_, "N",    0
     ) %>%
       mutate(
-        ADT = ymd(ADTC),
+        ADT = lubridate::ymd(ADTC),
         STUDYID = "XX1234",
         PARAMCD = "CRSP",
         PARAM = "Confirmed Response by Investigator"
@@ -183,7 +180,7 @@ test_that("derive_param_confirmed_resp Test 2: accept SD", {
 
 ## Test 3: error if invalid response values ----
 test_that("derive_param_confirmed_resp Test 3: error if invalid response values", {
-  adrs <- tribble(
+  adrs <- tibble::tribble(
     ~USUBJID, ~ADTC,        ~AVALC,
     "1",      "2020-01-01", "PR",
     "1",      "2020-02-01", "CR",
@@ -193,7 +190,7 @@ test_that("derive_param_confirmed_resp Test 3: error if invalid response values"
   ) %>%
     mutate(
       PARAMCD = "OVR",
-      ADT = ymd(ADTC),
+      ADT = lubridate::ymd(ADTC),
       STUDYID = "XX1234"
     )
 
@@ -221,7 +218,7 @@ test_that("derive_param_confirmed_resp Test 4: No source_pd", {
   #    - USUBJID == 1 will be a response as responsed after PD.
   #    - USUBJID == 2 will not be a responses as it straddles a PD.
   #+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-  adrs <- tribble(
+  adrs <- tibble::tribble(
     ~USUBJID, ~ADTC,        ~AVALC,
     "1",      "2020-01-01", "SD",
     "1",      "2020-02-01", "PD",
@@ -234,7 +231,7 @@ test_that("derive_param_confirmed_resp Test 4: No source_pd", {
   ) %>%
     mutate(
       PARAMCD = "OVR",
-      ADT = ymd(ADTC),
+      ADT = lubridate::ymd(ADTC),
       STUDYID = "XX1234"
     )
 
@@ -247,10 +244,76 @@ test_that("derive_param_confirmed_resp Test 4: No source_pd", {
       source_datasets = NULL,
       ref_confirm = 28,
       set_values_to = exprs(
+        AVAL = yn_to_numeric(AVALC),
         PARAMCD = "CRSP",
         PARAM = "Confirmed Response by Investigator"
       )
     )
+
+  expected_no_source_pd <- bind_rows(
+    adrs,
+    tribble(
+      ~USUBJID, ~ADTC,         ~AVALC, ~AVAL,
+      "1",      "2020-02-16",  "Y",    1,
+      "2",      NA_character_, "N",    0,
+      "3",      NA_character_, "N",    0,
+      "4",      NA_character_, "N",    0,
+      "5",      NA_character_, "N",    0,
+      "6",      NA_character_, "N",    0,
+      "7",      NA_character_, "N",    0,
+      "8",      NA_character_, "N",    0,
+      "9",      NA_character_, "N",    0
+    ) %>%
+      mutate(
+        ADT = ymd(ADTC),
+        STUDYID = "XX1234",
+        PARAMCD = "CRSP",
+        PARAM = "Confirmed Response by Investigator"
+      )
+  )
+
+  expect_dfs_equal(
+    base = expected_no_source_pd,
+    compare = actual_no_source_pd,
+    keys = c("USUBJID", "PARAMCD", "ADT")
+  )
+})
+
+test_that("derive_param_confirmed_resp Test 4: Deprecation warning for aval_fun", {
+  adrs <- tibble::tribble(
+    ~USUBJID, ~ADTC,        ~AVALC,
+    "1",      "2020-01-01", "SD",
+    "1",      "2020-02-01", "PD",
+    "1",      "2020-02-16", "PR",
+    "1",      "2020-03-21", "CR",
+    "2",      "2020-01-01", "SD",
+    "2",      "2020-02-01", "CR",
+    "2",      "2020-02-16", "PD",
+    "2",      "2020-03-21", "CR",
+  ) %>%
+    mutate(
+      PARAMCD = "OVR",
+      ADT = lubridate::ymd(ADTC),
+      STUDYID = "XX1234"
+    )
+
+  expect_warning(
+  actual_no_source_pd <-
+    derive_param_confirmed_resp(
+      adrs,
+      dataset_adsl = adsl,
+      filter_source = PARAMCD == "OVR",
+      source_pd = NULL,
+      source_datasets = NULL,
+      ref_confirm = 28,
+      aval_fun = yn_to_numeric,
+      set_values_to = exprs(
+        PARAMCD = "CRSP",
+        PARAM = "Confirmed Response by Investigator"
+      )
+    ),
+  class = "lifecycle_warning_deprecated"
+  )
 
   expected_no_source_pd <- bind_rows(
     adrs,