+++ tests/testthat/test-derive_param_response.R
@@ -1,22 +1,4 @@
-########################################################################
-#  Description: Unit Testing Exposed Function derive_param_response    #
-########################################################################
-
-#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-# Load required packages ----
-#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-
-library(tibble)
-library(lubridate)
-library(dplyr)
-library(admiraldev)
-library(admiral)
-
-#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-# Create Test Data ----
-#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-
-adsl <- tribble(
+adsl <- tibble::tribble(
   ~USUBJID, ~TRTSDTC,      ~CHECKKEPTCOL, ~CHECKNOTKEPTCOL,
   "1",      "2020-01-01",  "001",         "991",
   "2",      "2019-12-12",  "002",         "992",
@@ -24,37 +6,37 @@ adsl <- tribble(
   "4",      "2019-12-30",  "004",         "994",
 ) %>%
   mutate(
-    TRTSDT = ymd(TRTSDTC),
+    TRTSDT = lubridate::ymd(TRTSDTC),
     STUDYID = "XX1234"
   )
 
-adrs <- tribble(
-  ~USUBJID, ~ADTC, ~PARAMCD, ~AVALC, ~CHECKKEPTCOL,
-  "1", "2020-01-02", "OVR", "PR", "001",
-  "1", "2020-02-01", "OVR", "CR", "001",
-  "1", "2020-03-01", "OVR", "CR", "001",
-  "1", "2020-04-01", "OVR", "SD", "001",
-  "1", NA_character_, "PD", "N", "001",
-  "2", "2021-06-15", "OVR", "SD", "002",
-  "2", "2021-07-16", "OVR", "PD", "002",
-  "2", "2021-09-14", "OVR", "PD", "002",
-  "2", "2021-09-14", "PD", "Y", "002",
-  "3", "2021-09-14", "OVR", "SD", "003",
-  "3", "2021-10-30", "OVR", "PD", "003",
-  "3", "2021-12-25", "OVR", "CR", "003",
-  "3", "2021-10-30", "PD", "Y", "003"
+adrs <- tibble::tribble(
+  ~USUBJID, ~ADTC,        ~PARAMCD, ~AVALC, ~CHECKKEPTCOL,
+  "1",      "2020-01-02", "OVR",    "PR",   "001",
+  "1",      "2020-02-01", "OVR",    "CR",   "001",
+  "1",      "2020-03-01", "OVR",    "CR",   "001",
+  "1",      "2020-04-01", "OVR",    "SD",   "001",
+  "1",      NA,           "PD",     "N",    "001",
+  "2",      "2021-06-15", "OVR",    "SD",   "002",
+  "2",      "2021-07-16", "OVR",    "PD",   "002",
+  "2",      "2021-09-14", "OVR",    "PD",   "002",
+  "2",      "2021-09-14", "PD",     "Y",    "002",
+  "3",      "2021-09-14", "OVR",    "SD",   "003",
+  "3",      "2021-10-30", "OVR",    "PD",   "003",
+  "3",      "2021-12-25", "OVR",    "CR",   "003",
+  "3",      "2021-10-30", "PD",     "Y",    "003"
 ) %>%
   mutate(
     STUDYID = "XX1234",
-    ADT = ymd(ADTC),
+    ADT = lubridate::ymd(ADTC),
   ) %>%
   select(-ADTC)
 
-# derive_param_response, Test 1 ----
-test_that("Test 1: Test that response is derived accurately, with source_pd", {
+## Test 1: With source_pd ----
+test_that("derive_param_response Test 1: With source_pd", {
   expected_output <- bind_rows(
     adrs,
-    tribble(
+    tibble::tribble(
       ~USUBJID, ~ADTC,        ~AVALC, ~AVAL, ~CHECKKEPTCOL,
       "1",      "2020-01-02", "Y",    1,     "001",
       "2",      "",           "N",    0,     "002",
@@ -62,7 +44,7 @@ test_that("Test 1: Test that response is derived accurately, with source_pd", {
       "4",      "",           "N",    0,     "004",
     ) %>%
       mutate(
-        ADT = ymd(ADTC),
+        ADT = lubridate::ymd(ADTC),
         STUDYID = "XX1234",
         PARAMCD = "RSP",
         PARAM = "Response by investigator"
@@ -71,7 +53,7 @@ test_that("Test 1: Test that response is derived accurately, with source_pd", {
   )
 
   # Define the end of the assessment period for responses as the first PD date.
-  pd <- date_source(
+  pd <- admiral::date_source(
     dataset_name = "adrs",
     date = ADT,
     filter = PARAMCD == "PD" & AVALC == "Y"
@@ -85,6 +67,7 @@ test_that("Test 1: Test that response is derived accurately, with source_pd", {
       source_pd = pd,
       source_datasets = list(adrs = adrs),
       set_values_to = exprs(
+        AVAL = yn_to_numeric(AVALC),
         PARAMCD = "RSP",
         PARAM = "Response by investigator"
       ),
@@ -98,15 +81,11 @@ test_that("Test 1: Test that response is derived accurately, with source_pd", {
   )
 })
 
-# derive_param_response, Test 2 ----
-test_that("Test 2: Test that response is derived accurately, with No source_pd", {
-  #+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-  # source_pd = NULL, so the response on 2021-12-25 for subjid 3 is selected
-  #+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-
+## Test 2: No source_pd ----
+test_that("derive_param_response Test 2: No source_pd", {
   expected_output_no_source_pd <- bind_rows(
     adrs,
-    tribble(
+    tibble::tribble(
       ~USUBJID, ~ADTC,        ~AVALC, ~AVAL, ~CHECKKEPTCOL,
       "1",      "2020-01-02", "Y",    1,     "001",
       "2",      "",           "N",    0,     "002",
@@ -114,7 +93,7 @@ test_that("Test 2: Test that response is derived accurately, with No source_pd",
       "4",      "",           "N",    0,     "004",
     ) %>%
       mutate(
-        ADT = ymd(ADTC),
+        ADT = lubridate::ymd(ADTC),
         STUDYID = "XX1234",
         PARAMCD = "RSP",
         PARAM = "Response by investigator"
@@ -130,11 +109,57 @@ test_that("Test 2: Test that response is derived accurately, with No source_pd",
       source_pd = NULL,
       source_datasets = NULL,
       set_values_to = exprs(
+        AVAL = yn_to_numeric(AVALC),
         PARAMCD = "RSP",
         PARAM = "Response by investigator"
       ),
       subject_keys = get_admiral_option("subject_keys")
     )
+
+  expect_dfs_equal(
+    actual_output_no_source_pd,
+    expected_output_no_source_pd,
+    keys = c("USUBJID", "PARAMCD", "ADT")
+  )
+})
+
+## Test 3: deprecation warning for aval_fun ----
+test_that("derive_param_response Test 3: deprecation warning for aval_fun", {
+  expected_output_no_source_pd <- bind_rows(
+    adrs,
+    tibble::tribble(
+      ~USUBJID, ~ADTC,        ~AVALC, ~AVAL, ~CHECKKEPTCOL,
+      "1",      "2020-01-02", "Y",    1,     "001",
+      "2",      "",           "N",    0,     "002",
+      "3",      "2021-12-25", "Y",    1,     "003",
+      "4",      "",           "N",    0,     "004",
+    ) %>%
+      mutate(
+        ADT = lubridate::ymd(ADTC),
+        STUDYID = "XX1234",
+        PARAMCD = "RSP",
+        PARAM = "Response by investigator"
+      ) %>%
+      select(-ADTC)
+  )
+
+  # Derive the response parameter
+  expect_warning(
+  actual_output_no_source_pd <- adrs %>%
+    derive_param_response(
+      dataset_adsl = adsl,
+      filter_source = PARAMCD == "OVR" & AVALC %in% c("CR", "PR"),
+      source_pd = NULL,
+      source_datasets = NULL,
+      aval_fun = yn_to_numeric,
+      set_values_to = exprs(
+        PARAMCD = "RSP",
+        PARAM = "Response by investigator"
+      ),
+      subject_keys = get_admiral_option("subject_keys")
+    ),
+  class = "lifecycle_warning_deprecated"
+  )
 
   expect_dfs_equal(
     actual_output_no_source_pd,