+++ tests/testthat/test-utils.R
@@ -17,11 +17,16 @@ test_that("call_aval_fun Test 1: AVAL is created", {
     )
   }
 
-  expect_dfs_equal(
-    call_aval_fun(
+  expect_warning(
+    actual <- call_aval_fun(
       data,
       yn_map
     ),
+    class = "lifecycle_warning_deprecated"
+  )
+
+  expect_dfs_equal(
+    actual,
     tribble(
       ~AVALC, ~AVAL,
       "YES",      1,
@@ -38,9 +43,12 @@ test_that("call_aval_fun Test 2: Test error for invalid aval_fun", {
   }
 
   expect_error(
-    call_aval_fun(
+    suppress_warning(
+      call_aval_fun(
       data,
       aval_fun = bad_fun
+    ),
+    regexpr = "deprecated"
     ),
     regexp = "Assigning new AVAL records with aval_fun"
   )